Refactor navigation components and update env page routing
Introduced `OrgNav` and `EnvNav` components to streamline navigation logic and reduce duplication. Adjusted the layout and page structure to incorporate these new navigation components, enhancing code maintainability and clarity. Updated `page.tsx` to redirect to the overview page by default.
